Dshop/pyproject.toml
[tool.poetry]
name = "dshop"
version = "0.1.0"
description = "Django E-shop project"
authors = ["PyMasters"]
readme = "../README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
python-decouple = "^3.8"
pillow = "^10.0.0"
lxml = "^4.6"
requests = "^2.31.0"
sorl-thumbnail = "^12.9.0"
stripe = "^8.2.0"
requests_mock = "^1.11.0"
charset_normalizer = "^3.3.0"
freezegun = "^1.2.2"
django-tinymce = "^3.6.1"
dj-shop-cart = "^7.0.3"
djangorestframework="3.14.0"
drf-spectacular="0.26.5"
django-filter = "^23.5"
sentry-sdk = "^1.39.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
